requirements.txt
accelerate==0.33.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
args==0.1.0
aria2==0.0.1b0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
braceexpand==0.1.7
bypy==1.8.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
cmake==3.30.1
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
datasets==2.20.0
defusedxml==0.7.1
diffusers==0.21.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
emoji==2.12.1
exceptiongroup==1.2.2
fairscale==0.4.13
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
gdown==5.2.0
gekko==1.2.1
h11==0.14.0
hpsv2==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.24.2
humanfriendly==10.0
idna==3.7
image-reward==1.5
imageio==2.35.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.9.1
modelscope==1.20.1
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
open_clip_torch==2.32.0
openai==1.60.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum==1.21.3
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.2
protobuf==3.20.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.8.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rouge==1.0.1
safetensors==0.4.3
scikit-image==0.25.2
scipy==1.13.1
sentencepiece==0.2.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stanza==1.4.2
supervision==0.22.0
sympy==1.13.1
tensorboardX==2.6.2.2
tifffile==2025.5.10
timm==0.6.13
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
torchaudio==2.2.2+cu121
torchmetrics==1.7.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.43.3
triton==2.2.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.0.3
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.12.13
zipp==3.19.2
